Rename wagtail/admin/locale.py to localization.py
Having a module called locale interferes with Django's makemessages / compilemessages commands :-(
