+++ wagtail/admin/localization.py
