+++ wagtail/admin/wagtail_hooks.py
@@ -6,7 +6,7 @@ from draftjs_exporter.dom import DOM
 
 import wagtail.admin.rich_text.editors.draftail.features as draftail_features
 from wagtail.admin.auth import user_has_any_page_permission
-from wagtail.admin.locale import get_available_admin_languages, get_available_admin_time_zones
+from wagtail.admin.localization import get_available_admin_languages, get_available_admin_time_zones
 from wagtail.admin.menu import MenuItem, SubmenuMenuItem, settings_menu
 from wagtail.admin.navigation import get_explorable_root_page
 from wagtail.admin.rich_text import (
